+++ app/Http/Controllers/LegalisasiController.php
@@ -45,10 +45,12 @@ class LegalisasiController extends Controller
             ->addColumn('action', function ($data) {
                
                 $action = '';
-                $action .= "<a href='" . route('legalisasi.detail', $data->id_legalisasi) ."' class='btn btn-icon btn-success'><i class='fa fa-eye'></i></a>&nbsp;";
-                $action .= "<a href='javascript:void(0)' class='btn btn-icon btn-primary' data-id='{$data->id_legalisasi}' onclick='showLegalisasi(this);'><i class='fa fa-edit'></i></a>&nbsp;";
-                $action .= "<a href='javascript:void(0)' class='btn btn-icon btn-danger'  data-id='{$data->id_legalisasi}' onclick='deleteLegalisasi(this);'><i class='fa fa-trash'></i></a>&nbsp;";
-
+                $id_berkas = $data->berkas->id_user;
+                if ($id_berkas ==  Auth::user()->id_user ||  Auth::user()->level_user == 'Admin' || Auth::user()->level_user == 'Superadmin') {
+                    $action .= "<a href='" . route('legalisasi.detail', $data->id_legalisasi) ."' class='btn btn-icon btn-success'><i class='fa fa-eye'></i></a>&nbsp;";
+                    $action .= "<a href='javascript:void(0)' class='btn btn-icon btn-primary' data-id='{$data->id_legalisasi}' onclick='showLegalisasi(this);'><i class='fa fa-edit'></i></a>&nbsp;";
+                    $action .= "<a href='javascript:void(0)' class='btn btn-icon btn-danger'  data-id='{$data->id_legalisasi}' onclick='deleteLegalisasi(this);'><i class='fa fa-trash'></i></a>&nbsp;";
+                }
                 return $action;
             })
             ->addColumn('tanggal', function ($data) {
